pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foldseek-pocket-miner"
version = "0.1.0"
description = "A tool to find similar protein structures using Foldseek and generate PyMOL sessions with ligand pharmacophores and sequence conservation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Jinglin Jian", email = "jjian@scripps.edu"}
]
keywords = ["foldseek", "protein structure", "ligand", "pharmacophore", "pymol", "structural bioinformatics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "requests>=2.28.0",
    "biopython>=1.80",
    "numpy>=1.21.0",
    "pandas>=1.4.0",
    "tqdm>=4.64.0",
    "pymol-open-source>=2.5.0;platform_system!='Darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
]

[project.scripts]
foldseek-pocket-miner = "foldseek_pocket_miner.cli:main"

[project.urls]
Homepage = "https://github.com/JianJinglin/foldseek-pocket-miner"
Repository = "https://github.com/JianJinglin/foldseek-pocket-miner"
Issues = "https://github.com/JianJinglin/foldseek-pocket-miner/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
